% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_trajets}
\alias{compter_nombre_trajets}
\title{Compter le nombre de trajets}
\usage{
compter_nombre_trajets(trajet)
}
\arguments{
\item{trajet}{tibble. Données de trajets}
}
\value{
tibble. Données de comptage.
}
\description{
Compter le nombre de trajets
}
