% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_df_velo.R
\docType{data}
\name{df_velo}
\alias{df_velo}
\title{df_velo}
\format{
A data frame with 960 rows and 32 variables:
\describe{
\item{ Numéro de boucle }{  numeric }
\item{ Jour }{  Date }
\item{ 00 }{  numeric }
\item{ 01 }{  numeric }
\item{ 02 }{  numeric }
\item{ 03 }{  numeric }
\item{ 04 }{  numeric }
\item{ 05 }{  numeric }
\item{ 06 }{  numeric }
\item{ 07 }{  numeric }
\item{ 08 }{  numeric }
\item{ 09 }{  numeric }
\item{ 10 }{  numeric }
\item{ 11 }{  numeric }
\item{ 12 }{  numeric }
\item{ 13 }{  numeric }
\item{ 14 }{  numeric }
\item{ 15 }{  numeric }
\item{ 16 }{  numeric }
\item{ 17 }{  numeric }
\item{ 18 }{  numeric }
\item{ 19 }{  numeric }
\item{ 20 }{  numeric }
\item{ 21 }{  numeric }
\item{ 22 }{  numeric }
\item{ 23 }{  numeric }
\item{ Total }{  numeric }
\item{ Probabilité de présence d'anomalies }{  character }
\item{ Jour de la semaine }{  numeric }
\item{ Boucle de comptage }{  character }
\item{ Date formatée }{  Date }
\item{ Vacances }{  character }
}
}
\source{
https://data.nantesmetropole.fr/explore/dataset/244400404_comptages-velo-nantes-metropole/information/?disjunctive.boucle_num&disjunctive.jour_de_la_semaine&disjunctive.boucle_libelle&refine.jour=2025&refine.vacances_zone_b=Vacances+de+la+Toussaint
}
\usage{
df_velo
}
\description{
Données de vélos 2025 vacance de Toussaint.
}
\keyword{datasets}
