% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtre_anomalie}
\alias{filtre_anomalie}
\title{Filtre les anomalies}
\usage{
filtre_anomalie(trajet)
}
\arguments{
\item{trajet}{tibble. Données de trajets}
}
\value{
tibble. Données filtrées sans anomalies.
}
\description{
Filtre les anomalies
}
