% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtre_boucle}
\alias{filtre_boucle}
\title{Filtre les boucles par numéro}
\usage{
filtre_boucle(trajet, boucle)
}
\arguments{
\item{trajet}{tibble. Données de trajets}

\item{boucle}{character. Numéro de boucle à selectionner}
}
\value{
tibble. Données filtrées.
}
\description{
Filtre les boucles par numéro
}
\examples{
filtre_boucle(trajet = df_velo, boucle = c("880", "881"))
}
