% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{plot_distribution_semaine}
\alias{plot_distribution_semaine}
\title{Graph trjat par jour de la semaine}
\usage{
plot_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{}
}
\value{
ggplot. Graphique en barre
}
\description{
Graph trjat par jour de la semaine
}
