% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{trouver_trajet_max}
\alias{trouver_trajet_max}
\title{Trouver la paire boucle-jour avec le plus de trajet}
\usage{
trouver_trajet_max(trajet)
}
\arguments{
\item{trajet}{tibble. Données de trajets}
}
\value{
tibble. Données de la paires jour-boucle avec le plus de trajets.
}
\description{
Trouver la paire boucle-jour avec le plus de trajet
}
