% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{calcul_distribution_semaine}
\alias{calcul_distribution_semaine}
\title{Calcul du nombre de trajet par jour de la semaine}
\usage{
calcul_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{}
}
\value{
tibble. Données de comptage.
}
\description{
Calcul du nombre de trajet par jour de la semaine
}
